requirement.txt
Flask
joblib
numpy
pandas
python-dotenv
pymongo
